Move tensor constructors to aten
