+++ experimental/torch_xla2/test/test_functions.py
@@ -18,18 +18,6 @@ class TestTorchFunctions(parameterized.TestCase):
       ('tensor_empty', lambda: torch.tensor([],)),
       ('tensor_dtype', lambda: torch.tensor([[0.11111, 0.222222, 0.3333333]],
                                             dtype=torch.float64)),
-      ('ones_2d', lambda: torch.ones(2, 3)),
-      ('ones_1d', lambda: torch.ones(5)),
-      ('ones_1d_dtype', lambda: torch.ones(5, dtype=torch.float16)),
-      ('zeros_2d', lambda: torch.zeros(2, 3)),
-      ('zeros_1d', lambda: torch.zeros(5)),
-      ('zeros_1d_dtype', lambda: torch.zeros(5, dtype=torch.complex64)),
-      ('eye_3x3', lambda: torch.eye(3)),
-      ('eye_4x2', lambda: torch.eye(4, 2)),
-      ('eye_4x2_dtype', lambda: torch.eye(4, 2, dtype=torch.float16)),
-      ('full_2d', lambda: torch.full((2, 3), 3.141592)),
-      ('full_2d_dtype', lambda: torch.full(
-          (2, 3), 3.141592, dtype=torch.float16)),
   )
   def test_tensor_constructor(self, func: Callable[[], torch.Tensor]):
     expected = func()
